refactor manga list
